MCR-1917_MCR Move the method getSolrQuery from the MyCoRe module base to the SOLR module. So we get more independence between both modules.
